+++ client/src/features/onboarding/pages/top/components/wrapper/index.ts
@@ -1,0 +1,1 @@
+export * from "./wrapper";
