+++ client/src/features/onboarding/pages/top/components/wrapper/wrapper.css.ts
@@ -1,0 +1,11 @@
+import { style } from "@vanilla-extract/css";
+
+export const styles = {
+  container: style({
+    display: "flex",
+    alignItems: "center",
+    width: "100%",
+    height: "100vh",
+    justifyContent: "center",
+  }),
+};
