+++ client/src/features/onboarding/pages/top/components/wrapper/wrapper.tsx
@@ -1,0 +1,9 @@
+import { styles } from "./wrapper.css";
+
+type Props = {
+  children: React.ReactNode;
+};
+
+export const Wrapper = ({ children }: Props) => {
+  return <div className={styles.container}>{children}</div>;
+};
