+++ client/src/features/onboarding/pages/top/page.tsx
@@ -1,11 +1,12 @@
-import { styles } from "./page.css";
+import { Login } from "./components/login";
+import { Title } from "./components/title";
+import { Wrapper } from "./components/wrapper";
 
 export const OnboardingPage = () => {
   return (
-    <div className={styles.container}>
-      <img src="/maximum.svg" alt="Maximum" width="300" height="100" />
-      <h1 className={styles.title}>MELINE</h1>
-      <p className={styles.subtitle}>Coming Soon...</p>
-    </div>
+    <Wrapper>
+      <Title />
+      <Login />
+    </Wrapper>
   );
 };
